ST-8015 feat finalize helper file
