+++ src/currency/constants/currencySettings.js
@@ -1,0 +1,100 @@
+const currencySettings = {
+    MYR: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "RM",
+        signPosition: "BEFORE",
+    },
+    IDR: {
+        displayFormat: "0,0",
+        inputFormat: "0[.0000]",
+        sign: "Rp",
+        signPosition: "BEFORE",
+    },
+    HKD: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "HK$",
+        signPosition: "BEFORE",
+    },
+    SGD: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "$",
+        signPosition: "BEFORE",
+    },
+    THB: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "฿",
+        signPosition: "BEFORE",
+    },
+    VND: {
+        displayFormat: "0,0",
+        inputFormat: "0[.0000]",
+        sign: "₫",
+        signPosition: "AFTER",
+    },
+    PHP: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "₱",
+        signPosition: "BEFORE",
+    },
+    KHR: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "៛",
+        signPosition: "BEFORE",
+    },
+    AUD: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "$",
+        signPosition: "BEFORE",
+    },
+    USD: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "$",
+        signPosition: "BEFORE",
+    },
+    NZD: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "$",
+        signPosition: "BEFORE",
+    },
+    CNY: {
+        displayFormat: "0,0",
+        inputFormat: "0[.0000]",
+        sign: "¥",
+        signPosition: "BEFORE",
+    },
+    GBP: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "£",
+        signPosition: "BEFORE",
+    },
+    EUR: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "€",
+        signPosition: "BEFORE",
+    },
+    NOK: {
+        displayFormat: "0,0.00[00]",
+        inputFormat: "0.00[00]",
+        sign: "kr",
+        signPosition: "AFTER",
+    },
+    JPY: {
+        displayFormat: "0,0",
+        inputFormat: "0[.0000]",
+        sign: "¥",
+        signPosition: "BEFORE",
+    },
+};
+
+export default currencySettings;
