+++ src/currency/constants/index.js
@@ -1,2 +1,3 @@
 export { default as Currencies } from "./currencies";
 export { DefaultCurrency } from "./currencies";
+export { default as CurrencySettings } from "./currencySettings";
